review feedback + move stuff around
