+++ tools/src/versioning/android/packagesConfig.ts
